exclude rtcs without a recording from room streams upload completion check
its possible to have an rtc without a corresponding recording.
right now it fails room upload since we check if number of rtcs equals number of streams with a ready recording.
what we actually should be checking for is if number of rtcs equals number of streams with a ready recording OR without a recording